requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.1
openpyxl==3.1.2
pydantic==2.8.2
rapidfuzz==3.6.2
python-multipart==0.0.9
loguru==0.7.2
langchain-openai
langchain-google-genai
python-dotenv